We no longer need to move the debug python libs.
This was originally kept to be consistent with the move of the .dlls
to a common bin dir, but we are building static libs now.  Furthermore,
builds without vcpkg need to specify both the debug path and the
non-debug path to get other debug libs anyway, a mechanism which
continues to work for static Python as well.

For reference, the multiple paths should be used as follows:
> cmake -DCMAKE_PREFIX_PATH=$DEVLIBS;$DEVLIBS/debug [other parameters]